refactor(scrum-server): refactor UserType
